requirements.txt
xacro==1.13.3
pytest==6.2.4
deepdiff==5.5.0
pyyaml==5.3.1
xmlschema==1.6.4
